Implementação de leitura de header na requisição

+++ src/main/java/com/gabriel/martins/apinaescola/controller/EscolaController.java
@@ -12,6 +12,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
 
 @CrossOrigin
@@ -22,52 +23,72 @@ public class EscolaController {
     private EscolaService service;
     
     @PostMapping("/escola")
-    public ResponseEntity<?> cadastrarEscola(@RequestBody EscolaEntity escola){
-        try {
-            service.save(escola);
-            return ResponseEntity.status(HttpStatus.ACCEPTED).body(escola);
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Não foi possível cadastrar.");
+    public ResponseEntity<?> cadastrarEscola(@RequestHeader("api-hash") String hash, @RequestBody EscolaEntity escola){
+        if(hash.isBlank() || hash.isEmpty()){
+            try {
+                service.save(escola);
+                return ResponseEntity.status(HttpStatus.ACCEPTED).body(escola);
+            } catch (Exception e) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Não foi possível cadastrar.");
+            }
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Necessário autenticação.");
         }
     }
     
     @PostMapping("/escola/login")
-    public ResponseEntity<?> loginEscola(@RequestBody EscolaEntity escola){
-        try {
-            escola = service.findByLogin(escola);
-            return ResponseEntity.status(HttpStatus.ACCEPTED).body(escola);
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Usuário e/ou senha incorretos.");
+    public ResponseEntity<?> loginEscola(@RequestHeader("api-hash") String hash, @RequestBody EscolaEntity escola){
+        if(hash.isBlank() || hash.isEmpty()){
+            try {
+                escola = service.findByLogin(escola);
+                return ResponseEntity.status(HttpStatus.ACCEPTED).body(escola);
+            } catch (Exception e) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Usuário e/ou senha incorretos.");
+            }
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Necessário autenticação.");
         }
     }
     
     @GetMapping("/escola/{id}")
-    public ResponseEntity<?> findById(@PathVariable Long id) {
-        try {
-            EscolaEntity escola = service.findById(id);
-            return ResponseEntity.status(HttpStatus.ACCEPTED).body(escola);
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Necessário enviar um identificador válido.");
+    public ResponseEntity<?> findById(@RequestHeader("api-hash") String hash, @PathVariable Long id) {
+        if(hash.isBlank() || hash.isEmpty()){
+            try {
+                EscolaEntity escola = service.findById(id);
+                return ResponseEntity.status(HttpStatus.ACCEPTED).body(escola);
+            } catch (Exception e) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Necessário enviar um identificador válido.");
+            }
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Necessário autenticação.");
         }
     }
 
     @DeleteMapping("/escola/{id}")
-    public ResponseEntity<?> deleteUserById(@PathVariable Long id) {
-        try {
-            service.remove(id);
-            return ResponseEntity.status(HttpStatus.ACCEPTED).body(Boolean.TRUE);
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Necessário enviar um identificador válido.");
+    public ResponseEntity<?> deleteUserById(@RequestHeader("api-hash") String hash, @PathVariable Long id) {
+        if(hash.isBlank() || hash.isEmpty()){
+            try {
+                service.remove(id);
+                return ResponseEntity.status(HttpStatus.ACCEPTED).body(Boolean.TRUE);
+            } catch (Exception e) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Necessário enviar um identificador válido.");
+            }
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Necessário autenticação.");
         }
     }
 
     @PutMapping("/escola/{id}")
-    public ResponseEntity<?> updateUserById(@PathVariable Long id, @RequestBody EscolaEntity escola) {
-        try {
-            service.update(escola);
-            return ResponseEntity.status(HttpStatus.ACCEPTED).body(escola);
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Requisição inválida.");
+    public ResponseEntity<?> updateUserById(@RequestHeader("api-hash") String hash, @PathVariable Long id, @RequestBody EscolaEntity escola) {
+        if(hash.isBlank() || hash.isEmpty()){
+            try {
+                service.update(escola);
+                return ResponseEntity.status(HttpStatus.ACCEPTED).body(escola);
+            } catch (Exception e) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Requisição inválida.");
+            }
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Necessário autenticação.");
         }
     }
 }
